pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-extractor"
version = "0.1.0"
description = "Extract images from PDF, DOCX, and PPTX files"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "pillow>=9.0.0",
    "pypdf>=3.0.0",
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/image-extractor-cli"
"Bug Tracker" = "https://github.com/yourusername/image-extractor-cli/issues"

[project.scripts]
image-extractor = "image_extractor.cli:main"

[tool.setuptools]
packages = ["image_extractor"] 